Fix bug
Changed truncated to false and terminated to true for reaching max episode length. Also added a reward for meeting this criteria. Seems to mostly fix the index error when training.

Been playing around with hyperparameters, but not set on the best yet. Should probably run Tune. Agent is still converging to just avoid the obstacle and reach max episode.